+++ ImportApps/bkb_import_bank_statement_csv/test/ch.banana.switzerland.import.bkb.test.js
@@ -36,32 +36,34 @@ function TestImportBkb() {
 }
 
 // This method will be called at the beginning of the test case
-TestImportBkb.prototype.initTestCase = function() {
+TestImportBkb.prototype.initTestCase = function () {
    this.testLogger = Test.logger;
    this.progressBar = Banana.application.progressBar;
 }
 
 // This method will be called at the end of the test case
-TestImportBkb.prototype.cleanupTestCase = function() {
+TestImportBkb.prototype.cleanupTestCase = function () {
 
 }
 
 // This method will be called before every test method is executed
-TestImportBkb.prototype.init = function() {
+TestImportBkb.prototype.init = function () {
 
 }
 
 // This method will be called after every test method is executed
-TestImportBkb.prototype.cleanup = function() {
+TestImportBkb.prototype.cleanup = function () {
 
 }
 
-TestImportBkb.prototype.testImport = function() {
+TestImportBkb.prototype.testImport = function () {
    var fileNameList = [];
 
    fileNameList.push("file:script/../test/testcases/csv_bkb_example_format1A_20163006.csv");
    fileNameList.push("file:script/../test/testcases/csv_bkb_example_format1B_20230122.csv");
-   
+   fileNameList.push("file:script/../test/testcases/csv_bkb_example_format2_20240326.csv");
+   fileNameList.push("file:script/../test/testcases/csv_bkb_example_format2_20240327.csv");
+
    var parentLogger = this.testLogger;
    this.progressBar.start(fileNameList.length);
 
@@ -73,9 +75,9 @@ TestImportBkb.prototype.testImport = function() {
       Test.assert(file);
       var fileContent = file.read();
       Test.assert(fileContent);
-      var transactions = exec(fileContent,true); //takes the exec from the import script.
+      var transactions = exec(fileContent, true); //takes the exec from the import script.
       this.testLogger.addCsv('', transactions);
-      
+
       if (!this.progressBar.step())
          break;
    }