+++ ImportApps/credit_suisse_import_bank_statement_csv/test/ch.banana.switzerland.import.creditsuisse.test.js
@@ -66,6 +66,7 @@ TestImportCreditSuisse.prototype.testImport = function () {
    fileNameList.push("file:script/../test/testcases/csv_creditsuisse_example_format3_20171103.csv");
    fileNameList.push("file:script/../test/testcases/csv_creditsuisse_example_format4_20230901.csv");
    fileNameList.push("file:script/../test/testcases/csv_creditsuisse_example_format4_20230905.csv");
+   fileNameList.push("file:script/../test/testcases/csv_creditsuisse_example_format5_20240108.csv");
 
    var parentLogger = this.testLogger;
    this.progressBar.start(fileNameList.length);