+++ ImportApps/pf_import_bank_statement_csv/test/ch.banana.switzerland.import.postfinance.test.js
@@ -70,6 +70,12 @@ TestImportPostfinanceTrans.prototype.testImport = function () {
     fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format4_20230509.csv");
     fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format5_20230901.csv");
     fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format1_CreditCard_20230929.csv");
+    fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format6_20240215-DE.csv");
+    fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format6_20240222-IT.csv");
+    fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format6_20240226-FR.csv");
+    fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format6_20240228-FR-B.csv");
+    fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format6_20240228-EN.csv");
+    fileNameList.push("file:script/../test/testcases/csv_postfinance_example_format6_20240407-noenddate.csv");
 
     var parentLogger = this.testLogger;
     this.progressBar.start(fileNameList.length);