+++ ImportApps/iso20022_camt_v2/changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
+2024-05-15;
+ * [NEW] Added the option parameters to choose to import Credits transactions, Debits transactions or All the transactions.  
+
 2023-02-08; 
  * [] Released new verion for SPS 2022